Lint fixes and PR Cleanups
