+++ java/arcs/android/devtools/AndroidManifest.xml
