+++ javatests/arcs/android/e2e/testapp/BUILD
@@ -28,7 +28,7 @@ arcs_kt_android_library(
     deps = [
         ":animal_arcs",
         ":person_arcs",
-        "//java/arcs/android/devTools",
+        "//java/arcs/android/devtools",
         "//java/arcs/android/host",
         "//java/arcs/android/sdk/host",
         "//java/arcs/android/storage/database",
