+++ javatests/arcs/android/e2e/testapp/TestActivity.kt
@@ -17,13 +17,12 @@ import android.content.Intent
 import android.content.ServiceConnection
 import android.os.Bundle
 import android.os.IBinder
-import android.util.Log
 import android.view.View
 import android.widget.Button
 import android.widget.RadioButton
 import android.widget.TextView
 import androidx.appcompat.app.AppCompatActivity
-import arcs.android.devTools.DevToolsService
+import arcs.android.devtools.DevToolsService
 import arcs.android.host.AndroidManifestHostRegistry
 import arcs.core.allocator.Allocator
 import arcs.core.common.ArcId
@@ -71,19 +70,16 @@ class TestActivity : AppCompatActivity() {
     private var resurrectionArcId: ArcId? = null
 
     private var devToolsService: DevToolsService? = null
-    private var devToolsBound: Boolean = false;
 
     private val connection = object : ServiceConnection {
 
         override fun onServiceConnected(className: ComponentName, service: IBinder) {
-            // We've bound to LocalService, cast the IBinder and get LocalService instance
             val binder = service as DevToolsService.LocalBinder
             devToolsService = binder.getService()
-            devToolsBound = true
         }
 
         override fun onServiceDisconnected(arg0: ComponentName) {
-            devToolsBound = false
+            devToolsService = null
         }
     }
 
@@ -142,14 +138,10 @@ class TestActivity : AppCompatActivity() {
             }
         }
 
-        Log.d("TestActivity", "Starting intent")
         val devToolsIntent = Intent(
             this, DevToolsService::class.java
-        ).also { intent ->
-            bindService(intent, connection, Context.BIND_AUTO_CREATE)
-        }
-        //startForegroundService(devToolsIntent)
-        Log.d("TestActivity", "Intent Started")
+        )
+        bindService(devToolsIntent, connection, Context.BIND_AUTO_CREATE)
     }
 
     override fun onNewIntent(intent: Intent?) {
@@ -464,9 +456,7 @@ class TestActivity : AppCompatActivity() {
         result2 = result
         resultView1.text = "1: $result1"
         resultView2.text = "2: $result2"
-        if (devToolsBound) {
-            devToolsService?.send(result);
-        }
+        devToolsService?.send(result)
     }
 
     companion object {
